+++ src/Akeneo/Pim/Enrichment/Product/back/API/Command/UpsertProductCommand.php
@@ -4,8 +4,12 @@ declare(strict_types=1);
 
 namespace Akeneo\Pim\Enrichment\Product\API\Command;
 
+use Akeneo\Pim\Enrichment\Product\API\Command\UserIntent\Association\AssociationUserIntent;
+use Akeneo\Pim\Enrichment\Product\API\Command\UserIntent\Association\AssociationUserIntentCollection;
 use Akeneo\Pim\Enrichment\Product\API\Command\UserIntent\CategoryUserIntent;
 use Akeneo\Pim\Enrichment\Product\API\Command\UserIntent\FamilyUserIntent;
+use Akeneo\Pim\Enrichment\Product\API\Command\UserIntent\Groups\GroupUserIntent;
+use Akeneo\Pim\Enrichment\Product\API\Command\UserIntent\ParentUserIntent;
 use Akeneo\Pim\Enrichment\Product\API\Command\UserIntent\SetEnabled;
 use Akeneo\Pim\Enrichment\Product\API\Command\UserIntent\UserIntent;
 use Akeneo\Pim\Enrichment\Product\API\Command\UserIntent\ValueUserIntent;
@@ -28,10 +32,10 @@ final class UpsertProductCommand
         private mixed $identifierUserIntent = null,
         private ?FamilyUserIntent $familyUserIntent = null,
         private ?CategoryUserIntent $categoryUserIntent = null,
-        private mixed $parentUserIntent = null,
-        private mixed $groupsUserIntent = null,
+        private ?ParentUserIntent $parentUserIntent = null,
+        private ?GroupUserIntent $groupUserIntent = null,
         private ?SetEnabled $enabledUserIntent = null,
-        private mixed $associationsUserIntent = null,
+        private ?AssociationUserIntentCollection $associationUserIntent = null,
         private array $valueUserIntents = []
     ) {
         Assert::allImplementsInterface($this->valueUserIntents, ValueUserIntent::class);
@@ -44,11 +48,17 @@ final class UpsertProductCommand
     {
         $valueUserIntents = [];
         $categoryUserIntent = null;
+        $groupUserIntent = null;
         $enabledUserIntent = null;
         $familyUserIntent = null;
+        $parentUserIntent = null;
+        $associationUserIntents = \array_filter($userIntents, fn ($userIntent): bool => $userIntent instanceof AssociationUserIntent);
         foreach ($userIntents as $userIntent) {
             if ($userIntent instanceof ValueUserIntent) {
                 $valueUserIntents[] = $userIntent;
+            } elseif ($userIntent instanceof GroupUserIntent) {
+                Assert::null($groupUserIntent, "Only one groups intent can be passed to the command.");
+                $groupUserIntent = $userIntent;
             } elseif ($userIntent instanceof SetEnabled) {
                 Assert::null($enabledUserIntent, "Only one enabled intent can be passed to the command.");
                 $enabledUserIntent = $userIntent;
@@ -58,6 +68,9 @@ final class UpsertProductCommand
             } elseif ($userIntent instanceof CategoryUserIntent) {
                 Assert::null($categoryUserIntent, 'Only one category intent can be passed to the command.');
                 $categoryUserIntent = $userIntent;
+            } elseif ($userIntent instanceof ParentUserIntent) {
+                Assert::null($parentUserIntent, 'Only one parent intent can be passed to the command.');
+                $parentUserIntent = $userIntent;
             }
         }
 
@@ -66,7 +79,10 @@ final class UpsertProductCommand
             productIdentifier: $productIdentifier,
             familyUserIntent: $familyUserIntent,
             categoryUserIntent: $categoryUserIntent,
+            parentUserIntent: $parentUserIntent,
+            groupUserIntent: $groupUserIntent,
             enabledUserIntent: $enabledUserIntent,
+            associationUserIntent: empty($associationUserIntents) ? null : new AssociationUserIntentCollection($associationUserIntents),
             valueUserIntents: $valueUserIntents
         );
     }
@@ -99,8 +115,23 @@ final class UpsertProductCommand
         return $this->valueUserIntents;
     }
 
+    public function parentUserIntent(): ?ParentUserIntent
+    {
+        return $this->parentUserIntent;
+    }
+
+    public function groupUserIntent(): ?GroupUserIntent
+    {
+        return $this->groupUserIntent;
+    }
+
     public function enabledUserIntent(): ?SetEnabled
     {
         return $this->enabledUserIntent;
+    }
+
+    public function associationUserIntent(): ?AssociationUserIntentCollection
+    {
+        return $this->associationUserIntent;
     }
 }