+++ src/Akeneo/Pim/Enrichment/Product/back/API/Command/UserIntent/SetMultiSelectValue.php
@@ -2,15 +2,15 @@
 
 declare(strict_types=1);
 
-/**
- * @copyright 2022 Akeneo SAS (https://www.akeneo.com)
- * @license   http://opensource.org/licenses/osl-3.0.php  Open Software License (OSL 3.0)
- */
 namespace Akeneo\Pim\Enrichment\Product\API\Command\UserIntent;
 
 use Webmozart\Assert\Assert;
 
-class SetMultiSelectValue implements ValueUserIntent
+/**
+ * @copyright 2022 Akeneo SAS (https://www.akeneo.com)
+ * @license   http://opensource.org/licenses/osl-3.0.php  Open Software License (OSL 3.0)
+ */
+final class SetMultiSelectValue implements ValueUserIntent
 {
     /**
      * @param array<string> $values